# Change ImmutableSet.Builder to keep overflow hashtables and only deduplicate within table to improve performance