@bobyangyf 2018-08-08
This contains some bugs not caught in tests. Will submit some patches later with more tests.